Add files via upload
This project analyzes product data by focusing on monthly review ratings and yearly spending patterns to provide insights into product performance and customer engagement. The first goal is to calculate the average monthly review rating for each product based on a review table. For each month, represented numerically, the project calculates the average star rating for each product, rounding to two decimal places. This allows for a month-by-month assessment of product reception, sorted chronologically by month and organized by product ID for clarity.

The second objective is to identify the highest-grossing products in each category based on spending data from the year 2022. By filtering the product spending information for 2022, the project computes the total spend for each product in each category and then identifies the top two products with the highest spend within each category. This provides a clear picture of the top-performing products by revenue, offering insights into which products drove the most spending in their respective categories for that year.